# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job to Invoice API",
    "version": "0.1.0",
    "description": "Owner authentication, workspace setup, and workspace-isolated draft jobs."
  },
  "paths": {
...
    "/v1/jobs/{jobId}/changes": {
      "post": {
        "summary": "Create or return the current editing change-order draft for an eligible pre-invoice job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Editing change draft with server-calculated previous, change, and new totals" },
          "401": { "description": "Missing or invalid access token" },
          "403": { "description": "Suspended, locked, setup incomplete, or entitlement required" },
          "404": { "description": "Unknown or cross-tenant job" },
          "409": { "description": "Idempotency mismatch, pending approval, or accepted scope changed" },
          "422": { "description": "Job not eligible, extra body fields, or invalid job id" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}